fix: get our types in order
