+++ packages/react-baseline-inputs/src/IntegerInput.tsx
@@ -1,8 +1,6 @@
 import * as React from "react";
-import { CustomInputProps } from "./types";
 import { Field } from "./Field";
-
-export type IntegerInputProps = CustomInputProps<"input", number | null>;
+import { IntegerInputProps } from "./types";
 
 /**
  * An HTML `<input type="number" />`, but with the following benefits:
@@ -13,8 +11,6 @@ export type IntegerInputProps = CustomInputProps<"input", number | null>;
 export const IntegerInput: React.FC<IntegerInputProps> = ({
   value,
   onChange,
-  label = false,
-  wrap = true,
   ...props
 }) => {
   const handleChange = React.useCallback(
@@ -27,7 +23,6 @@ export const IntegerInput: React.FC<IntegerInputProps> = ({
 
   return (
     <Field
-      label={label}
       render={inputProps => (
         <input
           {...inputProps}
