+++ packages/react-baseline-inputs/src/Select.tsx
@@ -1,17 +1,6 @@
 import * as React from "react";
-import { CustomInputProps } from "./types";
 import { Field } from "./Field";
-
-export interface SelectOption extends React.HTMLProps<HTMLOptionElement> {
-  value: string;
-  label: string;
-  key?: any;
-}
-
-export interface SelectProps extends CustomInputProps<"select", string | null> {
-  placeholder?: string;
-  options: Array<SelectOption | string>;
-}
+import { SelectProps, SelectOption } from "./types";
 
 const getOptionProps = ({
   label: _label,
@@ -33,11 +22,9 @@ const getOptionProps = ({
  */
 export const Select: React.FC<SelectProps> = ({
   onChange,
-  options,
+  options = [],
   placeholder,
   value,
-  label = false,
-  wrap = true,
   ...props
 }) => {
   const handleChange = React.useCallback(
@@ -47,7 +34,6 @@ export const Select: React.FC<SelectProps> = ({
 
   return (
     <Field
-      label={label}
       render={inputProps => (
         <select
           {...inputProps}
