+++ packages/react-baseline-inputs/src/types.ts
@@ -1,19 +1,77 @@
-type ElementType = keyof JSX.IntrinsicElements;
+import * as React from "react";
 
-export type ExtraInputProps<T extends ElementType> = Omit<
-  Omit<JSX.IntrinsicElements[T], "value">,
-  "onChange"
-> &
-  CustomFieldProps;
+interface FieldInputProps {
+  id?: string;
+  inline?: boolean;
+  large?: boolean;
+  small?: boolean;
+  touched?: boolean;
+  success?: boolean;
+  error?: React.ReactNode;
+  help?: React.ReactNode;
 
-export type CustomInputProps<T extends ElementType, V> = ExtraInputProps<T> & {
+  label?: React.ReactNode;
+  wrapper?: boolean;
+
+  fieldClassName?: string;
+  fieldInlineClassName?: string;
+  fieldValidClassName?: string;
+  fieldInvalidClassName?: string;
+  fieldTouchedClassName?: string;
+  fieldLargeClassName?: string;
+  fieldSmallClassName?: string;
+
+  inputClassName?: string;
+  inputValidClassName?: string;
+  inputInvalidClassName?: string;
+  inputTouchedClassName?: string;
+  inputInlineClassName?: string;
+  inputLargeClassName?: string;
+  inputSmallClassName?: string;
+
+  labelClassName?: string;
+  labelInlineClassName?: string;
+  labelLargeClassName?: string;
+  labelSmallClassName?: string;
+
+  helpClassName?: string;
+  errorClassName?: string;
+}
+
+interface ValueProps<V> {
   value: V;
-  label?: string;
-  wrap?: boolean;
   onChange: (value: V) => void;
-};
+}
 
-export type CustomFieldProps = {
-  label?: string;
-  wrap?: boolean;
-};
+type HTMLProps<T> = Omit<React.HTMLProps<T>, "value" | "onChange">;
+type InputProps<V, E = HTMLInputElement> = FieldInputProps &
+  HTMLProps<E> &
+  ValueProps<V>;
+
+export interface FieldProps extends FieldInputProps {
+  render: (props: object) => React.ReactNode;
+}
+
+export type TextInputProps = InputProps<string | null>;
+export type IntegerInputProps = InputProps<number | null>;
+export type CheckboxProps = InputProps<boolean | null>;
+export type DateInputProps = InputProps<string | null>;
+export type DateTimeInputProps = InputProps<string | null>;
+export type FloatInputProps = InputProps<number | null>;
+export type TextAreaProps = InputProps<string | null, HTMLTextAreaElement>;
+
+export interface SelectOption extends React.HTMLProps<HTMLOptionElement> {
+  value: string;
+  label: string;
+  key?: any;
+}
+
+export interface SelectProps
+  extends InputProps<string | null, HTMLSelectElement> {
+  placeholder?: string;
+  options?: Array<SelectOption | string>;
+}
+
+export type FileInputProps = FieldInputProps & { value?: any } & (
+    | { multiple?: false | null | undefined; onChange: (value: File) => void }
+    | { multiple: true; onChange: (value: FileList) => void });
